+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [44.2.0]
-- [AlertView] Added `ButtonAlignment` property, that now supports `Underlying` and `Inline` button alignments
+- [AlertView] Added `ButtonAlignment` property that determines whether the buttons should be aligned underneath the title/description, or to the top right.
 
 ## [44.1.3] 
 - [ContextMenu][iOS] ContextMenu's with mode `Pressed` will now rebuild the menu when one of the items' properties changes.